add flask-caching to all endpoints
