+++ src/predict/api/helpers/utils.py
@@ -1,5 +1,5 @@
 import traceback
-from datetime import datetime, timedelta
+from datetime import datetime
 
 import pandas as pd
 import requests
@@ -9,7 +9,6 @@ from google.cloud import bigquery
 import geojson
 from app import cache
 from config.constants import connect_mongo, Config
-from models.predict import get_forecasts
 
 load_dotenv()
 db = connect_mongo()
@@ -19,6 +18,35 @@ def date_to_str(date: datetime):
     return date.isoformat()
 
 
+def heatmap_cache_key():
+    args = request.args
+    airqloud = args.get('airqloud')
+    page = args.get('page')
+    limit = args.get('limit')
+    return f'{airqloud}_{page}_{limit}'
+
+
+def weekly_forecasts_cache_key():
+    # create new var of current date and time as a string
+    current_date = datetime.now().strftime("%Y-%m-%d")
+    args = request.args
+    site_id = args.get('site_id')
+    return f'{current_date}_{site_id}'
+
+
+def geo_coordinates_cache_key():
+    key = (
+            "geo_coordinates:"
+            + str(round(float(request.args.get("latitude")), 6))
+            + ":"
+            + str(round(float(request.args.get("longitude")), 6))
+            + ":"
+            + str(request.args.get("distance_in_metres"))
+    )
+    return key
+
+
+@cache.memoize(timeout=Config.CACHE_TIMEOUT)
 def convert_to_geojson(data):
     """
     converts a list of predictions to geojson format
@@ -36,6 +64,7 @@ def convert_to_geojson(data):
     return geojson.FeatureCollection(features)
 
 
+@cache.memoize(timeout=Config.CACHE_TIMEOUT)
 def get_gp_predictions(airqloud=None, page=1, limit=500):
     """Returns PM 2.5 predictions for a particular airqloud name or id."""
 
@@ -64,22 +93,9 @@ def get_gp_predictions(airqloud=None, page=1, limit=500):
     total_count = len(predictions)
     paginated_results = predictions[(page - 1) * limit:page * limit]
     return paginated_results, total_count
-    # return predictions, total_count
 
 
-def geo_coordinates_cache_key():
-    key = (
-            "geo_coordinates:"
-            + str(round(float(request.args.get("latitude")), 6))
-            + ":"
-            + str(round(float(request.args.get("longitude")), 6))
-            + ":"
-            + str(request.args.get("distance_in_metres"))
-    )
-    return key
-
-
-@cache.memoize(timeout=3600)
+@cache.memoize(timeout=Config.CACHE_TIMEOUT)
 def get_health_tips() -> list[dict]:
     try:
         response = requests.get(
@@ -95,7 +111,7 @@ def get_health_tips() -> list[dict]:
         return []
 
 
-@cache.cached(timeout=3600, key_prefix=geo_coordinates_cache_key)
+@cache.memoize(timeout=Config.CACHE_TIMEOUT)
 def get_predictions_by_geo_coordinates(
         latitude: float, longitude: float, distance_in_metres: int
 ) -> dict:
@@ -122,6 +138,7 @@ def get_predictions_by_geo_coordinates(
     return data
 
 
+@cache.memoize(timeout=Config.CACHE_TIMEOUT)
 def get_forecasts_helper(db_name):
     """
     Helper function to get forecasts for a given site_id and db_name
@@ -150,3 +167,23 @@ def get_forecasts_helper(db_name):
         return data, 200
     else:
         return jsonify({"message": "Invalid request method", "success": False}), 400
+
+
+@cache.memoize(timeout=Config.CACHE_TIMEOUT)
+def get_forecasts(site_id, db_name):
+    """Retrieves forecasts from the database for the specified site_id."""
+    db = connect_mongo()
+    site_forecasts = list(db[db_name].find(
+        {'site_id': site_id}, {'_id': 0}).sort([('$natural', -1)]).limit(1))
+
+    results = []
+    if len(site_forecasts) > 0:
+        for i in range(0, len(site_forecasts[0]['pm2_5'])):
+            time = site_forecasts[0]['time'][i]
+            pm2_5 = site_forecasts[0]['pm2_5'][i]
+            health_tips = site_forecasts[0]['health_tips'][i]
+            result = {'time': time, 'pm2_5': pm2_5, 'health_tips': health_tips}
+            results.append(result)
+
+    formatted_results = {'forecasts': results}
+    return formatted_results
